fix(ci): satisfy restore repository shape gates
